neutron: fix Neutron calls issue
A previous attempt to reduce the number of Neutron calls was to
efficient: we were not able to get Neutron metadata in all cases!

For instance, we first get an update on the MAC attributes. The cache
entry was then set to PendingPortMetadata and the Neutron probe failed
to get metadata. Later, the Contrail probe set the attribute
Neutron/iface-mac which is used by Neutron. But since
PendingPortMetadata has been set, the Neutron call was not realized.

In this commit, before a call to Neutron, we set the cache entry of
port metadatas (which can be zero). We only do another Neutron call if
port metadatas have changed.

Change-Id: Ib873eac286111c7a8e04c84fecc7c77335042ea3