http: don't render topology.html
The go template format conflicts with vuejs. Both are using the {{ }}
notation to interpolate variables in the html.

Instead of rendering the html just for setting the service used we can
query the api to retrieve this information.

Also, we do not load sidebar components like captures when in agent mode.

Change-Id: I3fed286cbc812993952a163c3913437131c57555