ui: fix initial topology creation
Since the nodes are drawn in one pass on inital sync there are missing
attributes on the nodes (only the enter selection is evaluated).

On node creation we must set the id and add the correct css classes.

On node update we don't need to set the id again but we still refresh css
classes like before.

Change-Id: I2851b8e34f24bcfa91a77ec4b09fe4894c8d2e8a